+++ src/lib/filehandler.h
@@ -43,6 +43,7 @@ sds sds_getline(sds s, FILE *fp, size_t max, int *nread);
 sds sds_getfile(sds s, const char *file_path, size_t max, bool remove_newline, bool warn, int *nread);
 sds sds_getfile_from_fp(sds s, FILE *fp, size_t max, bool remove_newline, int *nread);
 
+bool create_tmp_file(const char *filepath);
 FILE *open_tmp_file(sds filepath);
 bool rename_tmp_file(FILE *fp, sds tmp_file, bool write_rc);
 bool write_data_to_file(const char *filepath, const char *data, size_t data_len);