+++ src/scripts/interface_caches.c
@@ -23,6 +23,35 @@
 #include <string.h>
 
 /**
+ * Creates a temporary file
+ * @param lua_vm lua instance
+ * @return number of elements pushed to lua stack
+ */
+int lua_caches_tmp_file(lua_State *lua_vm) {
+    int n = lua_gettop(lua_vm);
+    if (n != 1) {
+        MYMPD_LOG_ERROR(NULL, "Lua - caches_tmp_file: Invalid number of arguments");
+        lua_pop(lua_vm, n);
+        return luaL_error(lua_vm, "Invalid number of arguments");
+    }
+    const char *filename = lua_tostring(lua_vm, 1);
+    if (filename == NULL) {
+        MYMPD_LOG_ERROR(NULL, "Lua - caches_tmp_file: filename is NULL");
+        lua_pop(lua_vm, n);
+        return luaL_error(lua_vm, "filename is NULL");
+    }
+    lua_pop(lua_vm, n);
+
+    if (create_tmp_file(filename) == true) {
+        lua_pushnumber(lua_vm, 0);
+    }
+    else {
+        lua_pushnumber(lua_vm, 1);
+    }
+    return 1;
+}
+
+/**
  * Updates the timestamp of a file
  * @param lua_vm lua instance
  * @return number of elements pushed to lua stack
@@ -45,7 +74,9 @@ int lua_caches_update_mtime(lua_State *lua_vm) {
     if (update_mtime(filename) == true) {
         lua_pushnumber(lua_vm, 0);
     }
-    lua_pushnumber(lua_vm, 1);
+    else {
+        lua_pushnumber(lua_vm, 1);
+    }
     return 1;
 }